Documentation Bug Fixes
Looped audio is now fixed. Stereo is fixed. I finally know why ComponentType.ReadWrite<>() exists (Burst support). Hopefully these are the last framework changes before release!

+++ Packages/com.latios.latios-framework/Core/Core/Systems/DestroyEntitiesOnSceneChangeSystem.cs
@@ -31,6 +31,9 @@ namespace Latios.Systems
 
         private void RealUpdateOnSceneChange(Scene unloaded, Scene loaded)
         {
+            if (!Enabled)
+                return;
+
             latiosWorld.ResumeNextFrame();
             EntityManager.AddComponent<LatiosSceneChangeDummyTag>(EntityManager.UniversalQuery);
             EntityManager.DestroyEntity(m_destroyQuery);
